BUG/MINOR: quic: Lack of precision when computing K (cubic only cc)
K cubic variable is stored in ms. But it was a formula with the second as unit
for the window difference parameter which was used to compute K without
considering the loss of information. Then the result was converted in ms (K *= 1000).
This leaded to a lack of precision and multiples of 1000 as values.

To fix this, use the same formula but with the window difference in ms as parameter
passed to the cubic function and remove the conversion.

Must be backported as far as 2.6.

(cherry picked from commit a6d40e09f7c9ebb99d5f08630a6482fbb7d04d26)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 3b51c3db6f62e40b42ef537b6a3dabffad760d7c)
[wt: adj ctx]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 642bddc3eb5d0494b82430aa17ae1f5e633e02d3)
Signed-off-by: Willy Tarreau <w@1wt.eu>